requirements.txt
cffi>=1.14
charset-normalizer>=2.0.6
coverage>=6.0
cycler>=0.10.0
Deprecated>=1.2.13
Flask>=2.0.2
Flask_Cors>=3.0.0
gitdb>=4.0.7
GitPython>=3.1.20
gitrisky>=0.1.3
google-cloud-storage>=0.34.0
gunicorn>=19.9.0
idna>=3.2
kiwisolver>=1.3.1
matplotlib>=3.3.4
numpy>=1.19.5
Pillow>=8.3.2
pip>=21.2.4
pycparser>=2.20
PyGithub>=1.55
PyJWT>=2.1.0
PyNaCl>=1.4.0
pyparsing>=2.4.7
python-dateutil>=2.8.2
python-dotenv>=0.19.0
requests>=2.26.0
simplejson>=3.17
scipy>=1.5.4
setuptools>=57.0.0
six>=1.16.0
smmap>=4.0.0
typing-extensions>=3.10.0.2
urllib3>=1.26.7
wheel>=0.36.2
wrapt>=1.12.1

